requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
pydantic-settings>=2.0
groq>=0.4.0
pdfplumber>=0.10.0
python-docx>=1.0.0
python-multipart>=0.0.6
pyodbc>=5.0.0
sqlalchemy>=2.0.0
